requirements.txt
litellm>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
django>=5.0.0
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
httpx>=0.27.0
gunicorn>=21.0.0
gevent>=24.0.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
